Split SeleniumVersionCheck tests into two classes, style improvements
